moved UserService and added validators
